- Webpack vs parcel - Parcel - Installing React from npm with parcel - Extracting Component Module
