+++ .gitignore
@@ -1,0 +1,4 @@
+dist/
+node_modules/
+.cache/
+.vscode/
