+++ src/App.js
@@ -1,10 +1,6 @@
-const Pet = props => {
-	return React.createElement("div", {}, [
-		React.createElement("h1", {}, props.name),
-		React.createElement("h2", {}, props.animal),
-		React.createElement("h2", {}, props.breed)
-	]);
-};
+import React from 'react';
+import { render } from 'react-dom';
+import Pet from './Pet';
 
 class App extends React.Component {
 	handleTitleClick() {
@@ -37,7 +33,7 @@ class App extends React.Component {
 	}
 }
 
-ReactDOM.render(
+render(
 	React.createElement(App),
 	document.getElementById("root")
 );
