+++ src/index.html
@@ -9,8 +9,6 @@
   </head>
   <body>
     <div id="root">Nothing to render</div>
-    <script src="https://unpkg.com/react@16.4.1/umd/react.development.js"></script>
-    <script src="https://unpkg.com/react-dom@16.4.1/umd/react-dom.development.js"></script>
     <script src="./App.js"></script>
   </body>
 </html>
